+++ src/program/ui/InputEditorModel.h
@@ -181,6 +181,7 @@ private:
     
 signals:
     void inputSetChanged();
+    void stateLoaded();
 
 };